+++ CHANGELOG.md
@@ -9,13 +9,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `Solution`: new kwarg `balance_charge` to enable automatic charge balancing
+- `Solution`: add speciation support to the native engine via `phreeqpython`
+- `Solution`: add keyword argument to enable automatic charge balancing
+- `Salt`: class is now MSONable (i.e., serializable via `as_dict` / `from_dict`)
 - `Solution`: new properties `anions`, `cations`, `neutrals` provide easy access to subsets `components`.
 - `Solution`: improvements to `get_total_amount`.
 - `Solution`: new function `get_components_by_element` that lists all species associated with elements in specific
   oxidation states.
 - `Solution`: new properties `elements` and `chemical_system`, new function `get_el_amt_dict` to compute the total
   number of moles of each element present in the Solution.
+
+### Changed
+
+- Update `test_salt_ion_match` to `pytest` format and add additional tests
+- Update `test_bulk_properties` to `pytest` format
+
+### Removed
+
+- `Solution.list_salts()` has been removed. See `Solution.get_salt_dict()` for equivalent functionality
+- `salt_ion_match.generate_salt_list` and `identify_salt` have been removed. See `Solution.get_salt_dict()` and `Solution.get_salt()` for equivalent functionality.
 
 ### Fixed